requirements.txt
awsebcli==3.14.4
botocore==1.11.6
cement==2.8.2
certifi==2018.4.16
chardet==3.0.4
colorama==0.3.9
defusedxml==0.5.0
Django==2.0.7
django-allauth==0.36.0
django-filter==1.1.0
django-rest-auth==0.9.3
django-summernote==0.8.8.7
django-webpack-loader==0.6.0
django-widget-tweaks==1.4.2
djangorestframework==3.8.2
docutils==0.14
et-xmlfile==1.0.1
idna==2.7
jdcal==1.4
jmespath==0.9.3
mysqlclient==1.3.13
oauthlib==2.1.0
openpyxl==2.5.5
pathspec==0.5.5
Pillow==5.2.0
python-dateutil==2.7.3
python3-openid==3.1.0
pytz==2018.5
PyYAML==3.13
requests==2.19.1
requests-oauthlib==1.0.0
semantic-version==2.5.0
six==1.11.0
tabulate==0.7.5
termcolor==1.1.0
urllib3==1.23
